Move events to its own file
